+++ lending/RELEASES.md
@@ -99,3 +99,13 @@ Based on:
 - [typescript v0.0.15] lending
 ### Releases
 - [NPM v0.0.15] https://www.npmjs.com/package/ryan-lending/v/0.0.15 - lending
+
+## 2024-07-05 00:50:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.326.3 (2.359.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.16] lending
+### Releases
+- [NPM v0.0.16] https://www.npmjs.com/package/ryan-lending/v/0.0.16 - lending